requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.6.3            # via -r requirements.in
async-timeout==3.0.1      # via aiohttp
attrs==20.2.0             # via aiohttp
chardet==3.0.4            # via aiohttp
idna==2.10                # via yarl
multidict==4.7.6          # via aiohttp, yarl
python-dateutil==2.8.1    # via -r requirements.in
six==1.15.0               # via python-dateutil
yarl==1.5.1               # via aiohttp
